Pin build of cclyzer to a specific commit
The idea was to make it easier to track cclyzerpp updates by not
requiring a file edit, but that's only useful for test runs.
We can only make one release per commit in this repo anyways, so
if we aren't making commits to edit a pinned hash, we'd need to
make empty commits anyways. Might as well get the reproducibility too.

Obvious in retrospect, ah well.