Update pinned cclyzer with less conservative globals and split output
When analyzing (whole-program) applications, we can safely (?) assume
that there is no external code which will be linked in to the final
binary and cause values in global variables to escape.

When analyzing libraries, we can at least treat writes to internal-
or private-marked globals as non-escaping (assuming the global itself
does not escape).

Instead of a combined `mutated_or_escaped_globals` relation, cc2json
now outputs mutated and escaped globals (& functions) separately.